@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pymavryk"
-version = "3.10.3"
+version = "3.12.1"
 description = "Python toolkit for Mavryk"
 license = "MIT"
 authors = [
@@ -30,6 +30,7 @@ classifiers = [
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: Implementation :: CPython",
     "Topic :: Software Development :: Libraries :: Application Frameworks",
     "Typing :: Typed",
@@ -40,36 +41,36 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.12"
+python = ">=3.8.1,<3.13"
 attrs = ">=21.4.0"
 base58 = "^2.1.1"
 cattrs = ">=22.1.0"
 click = "^8.1.3"
-cryptography = "^41.0.0"
+cryptography = ">=42.0.4"
 deprecation = "^2.1.0"
 docker = "^6.0.0"
 fastecdsa = "^2.2.3"
 jsonschema = "^4.3.2"
-mnemonic = "^0.20"
+mnemonic = "^0.21"
 netstruct = "^1.1.2"
 notebook = "^6.5.6"
 ply = "^3.11"
-py-ecc = "^6.0.0"
+py-ecc = "^7.0.0"
 pysodium = "^0.7.10"
 python-dateutil = "^2.8.2"
-requests = "^2.28.2"
+# NOTE: https://github.com/docker/docker-py/issues/3256
+requests = ">=2.28.2,<2.32.0"
 secp256k1 = "^0.14.0"
 simplejson = "^3.17.6"
 strict-rfc3339 = "^0.7"
 tabulate = "^0.9.0"
 testcontainers = "^3.7.0"
 tqdm = "^4.62.3"
-setuptools = "^67.8.0"
+setuptools = "^69.0.0"
 simple-bson = "^0.0.3"
 
 [tool.poetry.dev-dependencies]
 black = "*"
-bump2version = "*"
 diff-cover = "*"
 isort = "*"
 mypy = "*"
@@ -101,12 +102,12 @@ force_single_line = true
 
 [tool.black]
 line-length = 120
-target-version = ['py38', 'py39', 'py310', 'py311']
+target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
 skip-string-normalization = true
 
 [tool.ruff]
 line-length = 120
-ignore = [
+lint.ignore = [
     "B904",
     "C417",
     "C901",
@@ -121,16 +122,16 @@ ignore = [
     "F811",
     "F821",
     "F841",
-    "R504",
-    "R505",
-    "R506",
-    "R507",
+    "RET504",
+    "RET505",
+    "RET506",
+    "RET507",
     "SIM102",
     "SIM108",
     "SIM114",
 ]
-target-version = "py311"
-extend-select = ["B", "C", "R", "SIM"]
+target-version = "py312"
+lint.extend-select = ["B", "C", "RET", "SIM"]
 
 [build-system]
 requires = ["poetry_core>=1.0.0"]
